add new config so that you can specify which data to index
We've decided that we don't need the raw `timer` data when we have the `timerData`
data too. This keeps the default as indexing everything but allows users
of the backend to specify the types that they want to index.

The only thing that users need to be careful of, is if they choose to
override the setting then they need to use any custom names they've given
each data type.